refactor: reorganiza estrutura do projeto seguindo arquitetura em camadas (Clean)
