+++ src/app.module.ts
@@ -1,9 +1,8 @@
 import { Module } from '@nestjs/common';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
-import { TypeOrmModule } from '@nestjs/typeorm';
 import { ConfigModule } from '@nestjs/config';
-import { UsuarioModule } from './usuario/usuario.module';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { UsuarioModule } from './interfaces/http/usuario/usuario.module';
+import { AppController } from './app.controller';
 
 @Module({
   imports: [
@@ -12,11 +11,12 @@ import { UsuarioModule } from './usuario/usuario.module';
       type: 'postgres',
       url: process.env.DATABASE_URL,
       autoLoadEntities: true,
-      synchronize: true,
+      synchronize: false,
+      migrations: ['dist/infrastructure/database/migrations/*.js'],
+      migrationsRun: true,
     }),
     UsuarioModule,
   ],
   controllers: [AppController],
-  providers: [AppService],
 })
 export class AppModule {}
