+++ src/application/usuario/dtos/login-usuario.dto.ts
@@ -1,0 +1,4 @@
+export class LoginUsuarioDto {
+  email: string;
+  senha: string;
+}
