--- src/usuario/dto/login-usuario.dto.ts
@@ -1,5 +1,0 @@
-export class LoginUsuarioDto {
-    email: string;
-    senha: string;
-  }
-
